update accumulator checkpoints, update chainparams
- Accumulator checkpoints set for zerocoin v2, zerocoin release of v1 and
  v2 coins works
- Added current checkpoints for main
- On regtest we do not need midas
  - int would allow to set max to
  GMT: Wednesday, 18. May 2033 03:33:19
  - changed int to int64_t
- changed RejectOldSporkKey to 1 second after zerocoin v2 block started
  changed from: 1553990400, GMT: Sunday, 31. March 2019 00:00:00
  changed to:   1550046138, GMT: Wednesday, 13. February 2019 08:22:18

+++ src/chainparams.h
@@ -124,7 +124,7 @@ public:
     CAmount InvalidAmountFiltered() const { return nInvalidAmountFiltered; };
 
     int MidasStartHeight() const { return nMidasStartHeight;   }
-    int MidasStartTime()   const { return nMidasStartTime;     }
+    int64_t MidasStartTime()   const { return nMidasStartTime;     }
     int DGWStartHeight()   const { return nDGWStartHeight;     }
     int DGWStartTime()     const { return nDGWStartTime;       }
 
@@ -178,7 +178,7 @@ protected:
     int nBlockZerocoinV2;
 
     int nMidasStartHeight;
-    int nMidasStartTime;
+    int64_t nMidasStartTime;
     int nDGWStartHeight;
     int nDGWStartTime;
 };
